renaming auth actions file to authentication
