+++ src/App.vue
@@ -8,7 +8,7 @@
 
 <script>
 import axios from './axios'
-import { AUTH_LOGOUT } from './store/actions/auth'
+import { AUTH_LOGOUT } from './store/actions/authentication'
 const appLayout = 'app'
 
 export default {
