+++ src/components/authentication/logout.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import { AUTH_LOGOUT } from '../../store/actions/auth'
+import { AUTH_LOGOUT } from '../../store/actions/authentication'
 
 export default {
   methods: {
