+++ src/store/index.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import auth from './modules/auth'
+import auth from './modules/authentication'
 Vue.use(Vuex)
 
 export default new Vuex.Store({
